requirements.txt
feedparser==6.0.10
python-dotenv==1.0.1
python-dateutil==2.8.2
requests==2.31.0
openai==1.12.0
httpx==0.27.0
